DAOPunks Grant Proposal 22-0117
PDF of the Grants Proposal that was put to vote on Snapshot by DAOPunk Community. The vote passed on Jan 21, 2022 with a vote of 146 'Yes' to 4 'No'. In the Snapshot vote a link to a Google Doc was provided as a reference. A PDF of the proposal was then uploaded to IPFS to ensure that the original document with no changed could be viewed.

IPFS CID: QmchHwEpXZ7za2LRbPYxuPHoAbL597jZmNy22gTLc3H7Eh
Snapshot Vote URL: https://snapshot.org/#/daopunks.eth/proposal/0x1ce3d7422f62b98097afeb9a2474b391e34f926e92a8c6e24454b4fbf9b1470d